// services/enums/errors.messages.enum.ts
export enum ErrorsMessagesEnum {
    STRING = "Veuillez entrer une chaine de charactères.",
    NUMBER = "Veuillez entrer un nombre entier.",
    EMAIL = "Veuillez entrer un format d'email correct.",
    PASSWORD = "Le format du mot de passe n'est pas bon.",
    DATE = "Veuillez sélectionner une date.",
    NB_MIN = "Le nombre minimum est ",
    NB_MAX = "Le nombre maximum est ",
    NB_RANGE = "Le nombre doit être entre ",
    STR_MIN = "Le nombre minimum de charactères est ",
    STR_MAX = "Le nombre maximum de charactères est ",
    STR_RANGE = "Le nombre de charactères doit se situer entre ",
    REQUIRED = "Ce champ est requis",
    SELECT = "Veuillez sélectionner un ",
    IS_IDENTIC = "Les champs ne sont pas identiques."
}